[TASK] Add chapter "Fluid"
This chapter explains the initial steps to convert the static template
files into Fluid template files.

A quick introduction to Fluid, the basic directory structure and purpose
of "Public/" and "Private/" directories are part of that. It then
explains how to build a page layout template under "Private/Templates/",
a partial ("Jumbotron") and the website layout file under
"Private/Layouts/".