[WIP][TASK] Add new chapter "Extension Installation"
This change adds a new chapter that describes the installation of the
site package extension (TYPO3 backend login, Extension Manager, how to
create initial pages and how to add a TypoScript template). It also
explains how to preview a page.

This chapter is work in progress and only contains the initial points as
bullet points.