Clearify changes made in Default.html
I was struggling in the part "The Page Layout File" (line 238) - because after removing tags in the Default.html, the Default.html was empty, but the text suggested me, that there should still be some html inside. So I was a little bit confused and started to search, if I missed a download or sth. like that. The text now should, hopefully, point out that using the origin index.html, renamed to Default.html, is empty after removing all tags. Nevertheless there could be the second way, if using a custom Default.html, than some html will be left, after removing the mentioned tags. Hope you know what I mean :-)